==> 0 - DataAcquisition.ipynb <==
# DATA ACQUISITION

# Import required library
import pandas as pd
import numpy as np
import os
import yfinance as yf
from datetime import datetime

In this notebook we are going to acquire all data that we need for our project. We are going to consider a period of ten years
from 2014 to 2023. The reference index will be the NAS100; here we are going to acquire its price history and the history of its
shares over the last 10 years (where possible).

url = 'https://en.wikipedia.org/wiki/Nasdaq-100'

# Read all tables from the webpage
tables = pd.read_html(url)

# Check all tables to find the right one
for i, table in enumerate(tables):
    print(f"Table {i}")
    print(table.head(), "\n")

# Let's start by considering the shares that are currently componing our interest index
tickers = tables[4]["Ticker"]
tickers

We are going to use data about added and removed security along the index history to get a list of all shares to be considered

# Load data
nasdaq_constituent = pd.read_json("nasdaq_constituent.json")
nasdaq_constituent

tickers_added = nasdaq_constituent["symbol"].unique().tolist()
tickers_removed = nasdaq_constituent["removedTicker"].unique().tolist()

S1 = set(tickers)
S2 = set(tickers_added)
S3 = set(tickers_removed)
S4 = S2.union(S3)
new_tickers = list(S4.difference(S1))

len(new_tickers)

tickers_list = tickers.to_list()
total_tickers = tickers_list + new_tickers

total_tickers.remove('')

len(total_tickers)

# Define the end date
end_date = datetime(2023, 12, 31)
# Define the start date (10 years before the end date)
start_date = datetime(2013, 12, 31)

# Dictionary to store the valid tickers with sufficient historical data
valid_tickers = {}

for ticker in total_tickers:
    # Download data
    df = yf.download(ticker, start = start_date.strftime('%Y-%m-%d'), end = end_date.strftime('%Y-%m-%d'), progress = True)

    df = df[['Close']]

    valid_tickers[ticker] = df

# Combine valid tickers into a single DataFrame
if valid_tickers:
    combined_data = pd.concat(valid_tickers, axis=1)


combined_data

using a specific library we retrieve the history of trading days for the considered market and look if we have the same number of days for the downloaded data:

import pandas_market_calendars as mcal

nas = mcal.get_calendar("NASDAQ")
trading_days = nas.valid_days(start_date = start_date, end_date = end_date)
trading_days

combined_data.columns = combined_data.columns.get_level_values(0)

Probably not all the downloaded data are going to be useable because of missing values.
Let's define a function to check if data for a stock respect that there are non-NaN after the first historical data present in the time series and if there is enough data to perform a rolling regression:

def is_valid_column(series):
    """checks whether a column is valid according to the specified conditions:"""
    # Verify if there are NaN values
    if series.isnull().any():
        # Find first index that is not Nan
        first_non_nan_index = series.first_valid_index()
        if first_non_nan_index is not None:
            # Check whether all values after the first non-NaN are also non-NaN
            return series[first_non_nan_index:].isnull().sum() == 0
        else:
            # If all values are NaN, consider the column as invalid
            return False
    return True

# Filter df columns
valid_columns = [col for col in combined_data.columns if is_valid_column(combined_data[col])]
df_filtered = combined_data[valid_columns]
df_filtered

Let's compute logarithmic returns for all selected stocks:

log_returns = np.log( df_filtered / df_filtered.shift(1))
log_returns

Finally, we download the data of the index chosen as reference and calculate the logarithmic returns:

nas100 = yf.download('^NDX', start = start_date, end = end_date)
nas100

# Compute log returns
nas100['log_returns'] = np.log(nas100.Close / nas100.Close.shift(1))
nas100 = nas100[["Close", "log_returns"]]
nas100

Save all acquired data in a csv format:

df_filtered.to_csv("Data/StocksPrices.csv")
log_returns.to_csv("Data/StocksReturns.csv")
nas100.to_csv("Data/nas100.csv")
