==> 1 - Nasdaq_Compositions.ipynb <==
# Reconstruction of the NASDAQ index composition in the years covered by the analysis

In order to conduct an analysis as accurate as possible, we needed to retrieve not only the time series of prices for the stocks that currently make up the NASDAQ index, but also the change in the index's composition over the past years.

import pandas as pd
import numpy as np
import os
import re
from datetime import datetime

Through the document *nasdaq_constituent.json*, we have reconstructed all stocks that joined or ceased to be part of the Index during the period we considered:

changes = pd.read_json('nasdaq_constituent.json')
changes.date.astype

The last update of the NASDAQ in the year 2013 was on 23 December, so we exclude previous updates:

changes = changes[changes.date >= '2013-12-01']
changes.shape

dates = changes["date"].unique().tolist()
len(dates)

url = 'https://en.wikipedia.org/wiki/Nasdaq-100'
# Read all tables from the webpage
tables = pd.read_html(url)
tickers = tables[4]["Ticker"].to_list()
len(tickers)

We can observe that the number of stocks exceeds by one; this is because the google share appears in two versions (with and without dividends).

We reconstructed the history of the index composition during the period considered and saved the data in a dataframe:

# Creating the DataFrame with the column named "components"
composition = pd.DataFrame()

for date in dates:

    change = changes[changes.date == date]
    added = change["symbol"].tolist()
    removed = change["removedTicker"].tolist()

    # Unique elements in 'a'
    unique_added = [a for a in added if a not in removed]

    # Unique elements in 'b'
    unique_removed = [r for r in removed if r not in added]

    print("\n", date)
    print(len(tickers))


    new_tickers = []

    for r in unique_removed:

        if (r != '') & (r not in tickers):

            print("add: ", r)
            tickers.append(r)

    new_tickers = tickers.copy()


    print(len(new_tickers))

    new_tickers_2 = []

    for a in unique_added:

        if (a != '') & (a in new_tickers):

            print("remove ",a)
            new_tickers.remove(a)

    new_tickers_2 = new_tickers.copy()

    print(len(new_tickers_2))

    if (len(added) != len(removed)):
        count += 1

    temp_df = pd.DataFrame({"components": [new_tickers_2]})
    temp_df.index = [date]

    tickers = new_tickers_2.copy()

    composition = pd.concat([composition, temp_df], axis = 0)

composition.iloc[::-1]

composition.to_csv('Data/nas_comps.csv')
