==> 2 -  logReturnsRegression.ipynb <==
# ROLLING REGRESSION OF *LOGARITHMIC RETURNS*

import pandas as pd
import numpy as np
from statsmodels.regression.rolling import RollingOLS
import statsmodels.api as sm
from pandas_datareader import data as pdr

We start by importing our data acquired in the previous notebooks and parsing dates:


nas100 = pd.read_csv("Data/nas100.csv")
nas100.Date = pd.to_datetime(nas100.Date)
nas100.set_index("Date", inplace = True)

log_returns = pd.read_csv("Data/StocksReturns.csv")
log_returns.Date = pd.to_datetime(log_returns.Date)
log_returns.set_index("Date", inplace = True)

log_returns.head()

dates = nas100.index
dates

We download the values of Fama and French Three Factor Model to extract the risk-free rate, useful for the following steps:

ff_factors = pdr.get_data_famafrench('F-F_Research_Data_Factors_daily', start = "2014-01-01", end = "2023-12-31")[0]
ff_factors = ff_factors / 100  # Convert to decimal
ff_factors

We calculated stocks and index returns net of the risk-free rate, as prescribed by the single index model used:

risk_free_rate = ff_factors['RF']
excess_log_returns = log_returns.subtract(risk_free_rate, axis = 0)
excess_market_log_returns  = nas100["log_returns"].subtract(risk_free_rate, axis = 0)

excess_market_log_returns.head()

tickers = log_returns.columns.tolist()
results = dict()
rolling_window = 180

Now we can execute a rolling regression of log-returns for all the stocks:

for ticker in tickers:
    df_returns = excess_log_returns[ticker].dropna()
    if not df_returns.empty:
        starting_index = df_returns.index[0]
        market_returns = excess_market_log_returns.loc[starting_index:].dropna()
        if not market_returns.empty:
            # Aligning market data indices with ticker returns data indices
            aligned_df = df_returns.align(market_returns, join = 'inner')
            if aligned_df[0].shape[0] > rolling_window:  # this ensures that there is enough data for the rolling window;
                try:
                    model = RollingOLS(aligned_df[0], sm.add_constant(aligned_df[1]), window = rolling_window)
                    result = model.fit()
                    results[ticker] = result
                except Exception as e:
                    print(f"Error for the ticker {ticker}: {e}")
            else:
                print(f"Not enough data for the rolling window for the ticker {ticker}")
                tickers.remove(ticker)
        else:
            print(f"Market data are blank for the ticker {ticker}")
    else:
        print(f"Returns data are empty for the ticker {ticker}")

We can now extract all the parameters of the rolling regression we are interested in:

# Step 5: Creating a dictionary to store the parameters
params = {
    'R_squared': {},
    'variance_stock': {},
    'mean_return_real': {},
    'alpha': {},
    'beta': {},
    'beta_squared_sigma_m_squared': {},
    'variance_residual': {},
}

# extracting parameters for each stock
for ticker in tickers:
    # Check if there are results for the ticker
    if ticker in results:
        # Rolling R_squared
        params['R_squared'][ticker] = results[ticker].rsquared

        # Uses logarithmic excess returns
        params['variance_stock'][ticker] = excess_log_returns[ticker].rolling(rolling_window).var()
        params['mean_return_real'][ticker] = excess_log_returns[ticker].rolling(rolling_window).mean()

        params['alpha'][ticker] = results[ticker].params["const"]
        params['beta'][ticker] = results[ticker].params[0]

        params['beta_squared_sigma_m_squared'][ticker] = results[ticker].ess / rolling_window
        params['variance_residual'][ticker] = results[ticker].ssr / rolling_window

# Concatenate the DataFrames, filling with NaN where necessary
params_df = pd.concat({k: pd.DataFrame(v) for k, v in params.items()}, axis = 1)

# Ensures that all tickers are present in the final DataFrame, filling in with NaN where data are missing
params_df = params_df.reindex(index = excess_log_returns.index)

params_df

print(results["ABNB"].pvalues[:, 0].shape)
print(results["AAPL"].pvalues[:, 0].shape)
print(type(results["ABNB"].pvalues[:, 0]))

We also extract the p-values of the alpha parameter in order to check its significance later:

params2 = {

    "alpha_significativity": {}
}

# Create a common index covering the entire date range
common_index = excess_log_returns.index

for ticker in tickers:
    if ticker in results:
        try:
            pvalues_alpha = pd.Series(results[ticker].pvalues[:, 0], index=results[ticker].model.data.row_labels)
            pvalues_alpha = pvalues_alpha.reindex(common_index)
            params2["alpha_significativity"][ticker] = pvalues_alpha
        except Exception as e:
            print(f"Error while processing the ticker {ticker}: {e}")
            params2["alpha_significativity"][ticker] = pd.Series(np.nan, index=common_index)

# Create a DataFrame for each key in params2
dfs = {k: pd.DataFrame(v) for k, v in params2.items()}

# Concatenate the DataFrames, filling with NaN where necessary
params_df2 = pd.concat(dfs, axis=1)

params_df2

df = pd.concat([params_df, params_df2], axis = 1)
df

df.to_csv("Data/parameters.csv", index = True)
